pyproject.toml
[project]
name = "codon_optimisation"
version = "0.0.1"
authors = [
    { name = "Suzanne Scott" }
]
description = "A command line tool for codon optimisation"
readme = "README.md"
requires-python = ">3.9"
dependencies = [
    "biopython >= 1.83",
    "typer >= 0.12.3"
]

[project.scripts]
co = "codon_optimisation.cli:app"

[project.optional-dependencies]
test = [
    "pytest >= 7.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = [
    "hatchling >= 1.26",
    ]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]
